refactor: Move article rows to new component
